+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased]
 ### Added
 - Prune and display changed files by X days (7, 30, 60, 90) from HEAD in the file explorer.
+- Restrict commits between a start and end date for pruned trees.
 
 ## [0.1.3] - 2020-10-31
 ### Added